Replace `reschedule` with new DSL endpoints and refactor task scheduler (#64)
There is bug in `Sync` which can allow two messages in the same
`SyncGroup` to be emitted in one order but processed in the opposite
order.

This PR removes the reschedule DSL endpoint and replaces it with two new
DSL enpoints
1. `Group`
2. `Pool`

`Group` extends the original functionality of `Sync` to allow for
`SyncGroup`s with a concurrency larger than 1.

`Pool` adds new functionality to allow the user to request that certain
reactions run in thread pools that are not the main thread or the
default thread pool. The `Always` DSL makes use of this new DSL endpoint
by creating a single-threaded thread pool for each distinct `Always`
reaction.

All tasks, regardless of which thread pool they are executed on, are now
scheduled for execution in the one scheduler. This means that the
priority queue that `Sync` uses has now been removed

---

There is still a bug which causes the test for `BaseClock` to SIGABRT on
Mac. Valgrind's DRD tool on linux flags something to do with Direct
emits and ReactionStatistics but the precise cause is still unknown.

However, this bug existed prior to this PR

---------

Co-authored-by: Trent Houliston <trent@houliston.me>