Refactor tests to use an event style and fix some found bugs (#70)
Currently, the tests are very haphazard, and they also have REQUIRE etc
statements within the reactions themselves. Catch2 isn't designed to be
used in a multithreaded environment like this.

This changes the tests to instead log events that happen and then at the
end check that the correct events happened in order.