Correct the executable path for osx (#170)
on osx, the executable path was using the size variable, however this
wasn't holding the number of bytes put in the buffer, it instead just is
used to output the required size if there was not enough. As a result,
the string had nulls in it. Later when it was concatenated with an
extension for traces, the nulls meant that the final path was just the
binary name which was overwritten.

since PATH_MAX is used as the buffer size, we don't need to look at the
space required as it should be PATH_MAX or less by definition.