Cache pool information on reactions to avoid finding it each time (#174)
Currently every single task that is added to the scheduler is required
to lock a pool mutex to find it's pool. There are too many mutex locks
on this path currently and it's obviously slowing down the system as a
whole.

While lock free queues will hopefully eventually solve the lower level
queue problems, this higher level mutex to select a pool can be avoided
by caching it on the reaction as once a reaction is created it will
always target the same pool.