Wake up chrono controller when RTF is zero (#182)
Currently if the RTF is zero, chronocontroller will remain waiting until
the RTF or the time changes. This is a problem if we ever want a chrono
controller task to run at the same time as `now()` and time is frozen.
This PR fixes this by allowing chronocontroller to wake up when RTF is
zero.